% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.whos}
\alias{as.whos}
\alias{as.whos.data.frame}
\alias{as.whos.data.table}
\title{Convert objects to whos}
\usage{
as.whos(x)

\method{as.whos}{data.table}(x)

\method{as.whos}{data.frame}(x)
}
\arguments{
\item{x}{Object of type \code{\link{data.table}} or \code{\link{data.frame}}.}
}
\description{
Convert objects to whos
}
\examples{
an.object <- "Containing all my stuff"
w <- as.data.frame(whos())
as.whos(w)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

