% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOpt}
\alias{default.options}
\alias{getOpt}
\title{Retrieves package options}
\usage{
getOpt(x)

default.options()
}
\arguments{
\item{x}{Option to retrieve.}
}
\description{
Retrieves package options
}
\examples{
# This shows how to modify a column presented by whos.
# The new function only reports the size of non-S4 objects
# to improve execution time.
opt <- default.options()
opt$columns$bytes <- function(x) if(isS4(x)) NA else object.size(x)
options(synesthesia = opt)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

