% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whos.options}
\alias{whos.options}
\title{Set default behavior of the whos function}
\usage{
whos.options(exclude, report.S4.size)
}
\arguments{
\item{exclude}{Objects to exclude from view. Can be a character vector of
names, or an environment, but not any regular expressions so far.}

\item{report.S4.size}{Calculating the size of S4 objects with
\code{\link{object.size}} can take an annoyingly long time (seconds), set
this option to \code{FALSE} to skip it and get quicker execution.}
}
\value{
Nothing. The values are stored as global options.
}
\description{
Set default behavior of the whos function
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

