% Generated by roxygen2 (4.0.2): do not edit by hand
\name{whos.set.mask}
\alias{whos.set.mask}
\title{Set a default exclusion mask for \code{\link{whos}}.}
\usage{
whos.set.mask(exclude, envir = globalenv())
}
\arguments{
\item{exclude}{List of object names. These will be hidden from view.
Defaults to all objects in \code{envir}.}

\item{envir}{Environment to work in.}
}
\value{
Nothing. The mask is stored as `whos.exclude' in the global option list.
}
\description{
Set a default exclusion mask for \code{\link{whos}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

