% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entry_view.R
\name{entry.view}
\alias{entry.view}
\title{Display vectors, lists or rows of a data frames in key-value-pairs.}
\usage{
entry.view(x, i = 1)
}
\arguments{
\item{x}{List or data frame.}

\item{i}{Row number to show. Press down/up to browse.}
}
\value{
Nothing.
}
\description{
Color coded according to class of contents.
}
\examples{
entry.view(Sys.getenv())
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

